+++ ios/docs/html/_c_m_user_account_result_8h.js
@@ -1,6 +1,20 @@
 var _c_m_user_account_result_8h =
 [
-    [ "CMUserAccountResult", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138ee", null ],
+    [ "CMUserAccountResult", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138ee", [
+      [ "CMUserAccountUnknownResult", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eeaf58308f9e2b28eed6ef1facdf1bb2cf5", null ],
+      [ "CMUserAccountLoginSucceeded", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea5fc2e7b1014b676800b515f6e3948e98", null ],
+      [ "CMUserAccountLogoutSucceeded", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eeaeecaa9ed84d13cdd5a0fecc6c9d9fbce", null ],
+      [ "CMUserAccountCreateSucceeded", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea2abab18343e3a9fc1ab18df2f6c90fe9", null ],
+      [ "CMUserAccountProfileUpdateSucceeded", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea62e3edfbefcf4d3a37c3d438f0e20bda", null ],
+      [ "CMUserAccountPasswordChangeSucceeded", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eeaa764680c2ab1a1588ad51aee2a09ac76", null ],
+      [ "CMUserAccountPasswordResetEmailSent", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea823a98b55b291436c33c1a5f9522c47c", null ],
+      [ "CMUserAccountCreateFailedInvalidRequest", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea5670a1fbb118996aab2a55087a745a23", null ],
+      [ "CMUserAccountProfileUpdateFailed", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eeabc796f14eba5f6525bcbf068c108bd4d", null ],
+      [ "CMUserAccountCreateFailedDuplicateAccount", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea206cb8d3de756a7918e0d05e217cd522", null ],
+      [ "CMUserAccountLoginFailedIncorrectCredentials", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea079196bcb074f6a9b8ccc0d13d264cfe", null ],
+      [ "CMUserAccountPasswordChangeFailedInvalidCredentials", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eea79e1efb15fa7977fb4707ce7e7d451fb", null ],
+      [ "CMUserAccountOperationFailedUnknownAccount", "_c_m_user_account_result_8h.html#aaf8d3c3117387947ca9c709691b138eeafb9cf6702ad9ea4ad59c6422625d0720", null ]
+    ] ],
     [ "CMUserAccountOperationFailed", "_c_m_user_account_result_8h.html#a5aa9fd83164b304babc79fba16715864", null ],
     [ "CMUserAccountOperationSuccessful", "_c_m_user_account_result_8h.html#af7bd08ac73677fd2e017cefaf8b876ae", null ]
 ];