+++ ios/ios/src/Web Services/CMWebService.m
@@ -38,7 +38,6 @@ NSString * const YAJLErrorKey = @"YAJLErrorKey";
 @interface CMWebService () {
     NSMutableDictionary *_responseTimes;
     __strong CMWebServiceUserAccountOperationCallback temporaryCallback;
-
 }
 @property (nonatomic, strong) NSString *apiUrl;
 - (NSURL *)constructTextUrlAtUserLevel:(BOOL)atUserLevel withKeys:(NSArray *)keys query:(NSString *)searchString pagingOptions:(CMPagingDescriptor *)paging sortingOptions:(CMSortDescriptor *)sorting withServerSideFunction:(CMServerFunction *)function extraParameters:(NSDictionary *)params;
@@ -352,7 +351,6 @@ NSString * const YAJLErrorKey = @"YAJLErrorKey";
     } errorHandler:errorHandler];
 }
 
-
 #pragma mark - Singly Proxy
 
 - (void)runSocialGraphGETQueryOnNetwork:(NSString *)network
@@ -534,9 +532,10 @@ NSString * const YAJLErrorKey = @"YAJLErrorKey";
     }];
 }
 
-// ***** Singly Social
-- (void)loginWithSocial:(CMUser *)user withService:(NSString *)service viewController:(UIViewController *)viewController params:(NSString *)params callback:(CMWebServiceUserAccountOperationCallback)callback {
+
+- (void)loginWithSocial:(CMUser *)user withService:(NSString *)service viewController:(UIViewController *)viewController params:(NSDictionary *)params callback:(CMWebServiceUserAccountOperationCallback)callback {
     CMSocialLoginViewController *loginViewController = [[CMSocialLoginViewController alloc] initForService:service appID:_appIdentifier apiKey:_appSecret user:user params:params];
+    loginViewController.modalPresentationStyle = UIModalPresentationFormSheet;
     loginViewController.delegate = self;
     temporaryCallback = callback;
     [viewController presentViewController:loginViewController animated:YES completion:NULL];
@@ -579,11 +578,14 @@ NSString * const YAJLErrorKey = @"YAJLErrorKey";
             default:
                 break;
         }
+        
+        if ([messages valueForKey:@"expires"] == [NSNull null]) {
+            resultCode = CMUserAccountLoginFailedIncorrectCredentials;
+        }
+        
         temporaryCallback(resultCode, messages);
     }]; 
 }
-
-// ***** /Singly Social
 
 - (void)saveUser:(CMUser *)user callback:(CMWebServiceUserAccountOperationCallback)callback {
     NSParameterAssert(user);